requirements.txt
Django>=4.2
Pillow>=10.0.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7
django-widget-tweaks>=1.4.12
python-decouple>=3.8
whitenoise>=6.0
